fix(battle): satisfy clippy and tokei for split merge updates
