+++ crates/souprune/src/app_state/battle/collision.rs
@@ -16,7 +16,19 @@ use bevy::prelude::*;
 use bevy_alight_motion::sdf_material::SdfMaterial;
 use bevy_tween::interpolation::EaseKind;
 use serde::{Deserialize, Serialize};
-use std::collections::VecDeque;
+
+mod animation;
+mod geometry;
+
+use self::animation::{
+    animate_battle_box_merge_system, animate_battle_box_split_system, finalize_merged_battle_box,
+    spawn_standalone_box, spawn_standalone_box_entity,
+};
+use self::geometry::{
+    choose_box_index_for_player, merge_boundaries, plan_merge_animation, resolve_boundary,
+    resolve_live_battle_box, retire_battle_box, retire_existing_battle_boxes_with_id,
+    select_box_id_for_player, split_rect_box,
+};
 
 type UiBattleBoxReadQuery<'w, 's> = Query<
     'w,
@@ -386,103 +398,6 @@ impl BattleBoxCandidate {
 
 // ─── Split / Merge Algorithms ───────────────────────────────────────
 
-/// Split a rectangular boundary along an axis.
-fn split_rect_box(
-    original: &BattleBoxBoundary,
-    axis: &SplitAxis,
-    split_pos: f32,
-    gap: f32,
-    gap_policy: GapPolicy,
-) -> (BattleBoxBoundary, BattleBoxBoundary) {
-    match axis {
-        SplitAxis::Vertical => {
-            let left_width = original.half_size.x + split_pos;
-            let right_width = original.half_size.x - split_pos;
-            let half_gap = gap / 2.0;
-            let min_x = original.center.x - original.half_size.x;
-            let max_x = original.center.x + original.half_size.x;
-
-            match gap_policy {
-                GapPolicy::Expands => {
-                    let left = BattleBoxBoundary {
-                        half_size: Vec2::new(left_width / 2.0, original.half_size.y),
-                        center: Vec2::new(
-                            original.center.x - original.half_size.x + left_width / 2.0 - half_gap,
-                            original.center.y,
-                        ),
-                    };
-                    let right = BattleBoxBoundary {
-                        half_size: Vec2::new(right_width / 2.0, original.half_size.y),
-                        center: Vec2::new(
-                            original.center.x + original.half_size.x - right_width / 2.0 + half_gap,
-                            original.center.y,
-                        ),
-                    };
-                    (left, right)
-                }
-                GapPolicy::Includes => {
-                    let total_width = original.half_size.x * 2.0;
-                    let scale = (total_width - gap) / total_width;
-                    let left_width_scaled = left_width * scale;
-                    let right_width_scaled = right_width * scale;
-
-                    let left = BattleBoxBoundary {
-                        half_size: Vec2::new(left_width_scaled / 2.0, original.half_size.y),
-                        center: Vec2::new(min_x + left_width_scaled / 2.0, original.center.y),
-                    };
-                    let right = BattleBoxBoundary {
-                        half_size: Vec2::new(right_width_scaled / 2.0, original.half_size.y),
-                        center: Vec2::new(max_x - right_width_scaled / 2.0, original.center.y),
-                    };
-                    (left, right)
-                }
-            }
-        }
-        SplitAxis::Horizontal => {
-            // Positive split_pos moves the split line upward, making the top box smaller
-            // and the bottom box larger.
-            let top_height = original.half_size.y - split_pos;
-            let bottom_height = original.half_size.y + split_pos;
-            let half_gap = gap / 2.0;
-            let min_y = original.center.y - original.half_size.y;
-            let max_y = original.center.y + original.half_size.y;
-
-            match gap_policy {
-                GapPolicy::Expands => {
-                    let top = BattleBoxBoundary {
-                        half_size: Vec2::new(original.half_size.x, top_height / 2.0),
-                        center: Vec2::new(original.center.x, max_y + half_gap - top_height / 2.0),
-                    };
-                    let bottom = BattleBoxBoundary {
-                        half_size: Vec2::new(original.half_size.x, bottom_height / 2.0),
-                        center: Vec2::new(
-                            original.center.x,
-                            min_y - half_gap + bottom_height / 2.0,
-                        ),
-                    };
-                    (top, bottom)
-                }
-                GapPolicy::Includes => {
-                    let total_height = original.half_size.y * 2.0;
-                    let scale = (total_height - gap) / total_height;
-                    let top_height_scaled = top_height * scale;
-                    let bottom_height_scaled = bottom_height * scale;
-
-                    let top = BattleBoxBoundary {
-                        half_size: Vec2::new(original.half_size.x, top_height_scaled / 2.0),
-                        center: Vec2::new(original.center.x, max_y - top_height_scaled / 2.0),
-                    };
-                    let bottom = BattleBoxBoundary {
-                        half_size: Vec2::new(original.half_size.x, bottom_height_scaled / 2.0),
-                        center: Vec2::new(original.center.x, min_y + bottom_height_scaled / 2.0),
-                    };
-                    (top, bottom)
-                }
-            }
-        }
-    }
-}
-
 #[derive(Debug, Clone)]
 struct MergeAnimationPlan {
     axis: SplitAxis,
@@ -490,651 +405,6 @@ struct MergeAnimationPlan {
     target_boundary_a: BattleBoxBoundary,
     target_boundary_b: BattleBoxBoundary,
     merged_boundary: BattleBoxBoundary,
-}
-
-fn lerp_boundary(start: &BattleBoxBoundary, end: &BattleBoxBoundary, t: f32) -> BattleBoxBoundary {
-    BattleBoxBoundary {
-        center: start.center.lerp(end.center, t),
-        half_size: start.half_size.lerp(end.half_size, t),
-    }
-}
-
-fn infer_merge_axis(a: &BattleBoxBoundary, b: &BattleBoxBoundary) -> Option<SplitAxis> {
-    const EPSILON: f32 = 0.001;
-
-    let overlap_x = (a.half_size.x + b.half_size.x) - (a.center.x - b.center.x).abs();
-    let overlap_y = (a.half_size.y + b.half_size.y) - (a.center.y - b.center.y).abs();
-
-    if overlap_x + EPSILON < overlap_y {
-        return Some(SplitAxis::Vertical);
-    }
-    if overlap_y + EPSILON < overlap_x {
-        return Some(SplitAxis::Horizontal);
-    }
-
-    let delta_x = (a.center.x - b.center.x).abs();
-    let delta_y = (a.center.y - b.center.y).abs();
-
-    if delta_y + EPSILON < delta_x {
-        Some(SplitAxis::Vertical)
-    } else if delta_x + EPSILON < delta_y {
-        Some(SplitAxis::Horizontal)
-    } else {
-        None
-    }
-}
-
-fn plan_merge_animation(
-    boundaries: &[BattleBoxBoundary; 2],
-    gap_policy: GapPolicy,
-) -> Option<MergeAnimationPlan> {
-    let axis = infer_merge_axis(&boundaries[0], &boundaries[1])?;
-
-    match axis {
-        SplitAxis::Vertical => {
-            let ordered_indices = if boundaries[0].center.x <= boundaries[1].center.x {
-                (0, 1)
-            } else {
-                (1, 0)
-            };
-            let left = &boundaries[ordered_indices.0];
-            let right = &boundaries[ordered_indices.1];
-
-            let left_min = left.center.x - left.half_size.x;
-            let left_max = left.center.x + left.half_size.x;
-            let right_min = right.center.x - right.half_size.x;
-            let right_max = right.center.x + right.half_size.x;
-
-            let (target_left, target_right) = match gap_policy {
-                GapPolicy::Expands => {
-                    let split_x = (left_max + right_min) * 0.5;
-                    (
-                        BattleBoxBoundary {
-                            half_size: left.half_size,
-                            center: Vec2::new(split_x - left.half_size.x, left.center.y),
-                        },
-                        BattleBoxBoundary {
-                            half_size: right.half_size,
-                            center: Vec2::new(split_x + right.half_size.x, right.center.y),
-                        },
-                    )
-                }
-                GapPolicy::Includes => {
-                    let outer_span = right_max - left_min;
-                    let combined_width = left.half_size.x * 2.0 + right.half_size.x * 2.0;
-                    if outer_span <= 0.0 || combined_width <= 0.0 {
-                        return None;
-                    }
-
-                    let scale = combined_width / outer_span;
-                    if scale <= f32::EPSILON {
-                        return None;
-                    }
-
-                    let target_left_width = (left.half_size.x * 2.0) / scale;
-                    let target_right_width = (right.half_size.x * 2.0) / scale;
-                    (
-                        BattleBoxBoundary {
-                            half_size: Vec2::new(target_left_width * 0.5, left.half_size.y),
-                            center: Vec2::new(left_min + target_left_width * 0.5, left.center.y),
-                        },
-                        BattleBoxBoundary {
-                            half_size: Vec2::new(target_right_width * 0.5, right.half_size.y),
-                            center: Vec2::new(right_max - target_right_width * 0.5, right.center.y),
-                        },
-                    )
-                }
-            };
-
-            let merged_boundary = merge_boundaries(&target_left, &target_right);
-            Some(MergeAnimationPlan {
-                axis,
-                ordered_indices,
-                target_boundary_a: target_left,
-                target_boundary_b: target_right,
-                merged_boundary,
-            })
-        }
-        SplitAxis::Horizontal => {
-            let ordered_indices = if boundaries[0].center.y >= boundaries[1].center.y {
-                (0, 1)
-            } else {
-                (1, 0)
-            };
-            let top = &boundaries[ordered_indices.0];
-            let bottom = &boundaries[ordered_indices.1];
-
-            let top_max = top.center.y + top.half_size.y;
-            let top_min = top.center.y - top.half_size.y;
-            let bottom_max = bottom.center.y + bottom.half_size.y;
-            let bottom_min = bottom.center.y - bottom.half_size.y;
-
-            let (target_top, target_bottom) = match gap_policy {
-                GapPolicy::Expands => {
-                    let split_y = (top_min + bottom_max) * 0.5;
-                    (
-                        BattleBoxBoundary {
-                            half_size: top.half_size,
-                            center: Vec2::new(top.center.x, split_y + top.half_size.y),
-                        },
-                        BattleBoxBoundary {
-                            half_size: bottom.half_size,
-                            center: Vec2::new(bottom.center.x, split_y - bottom.half_size.y),
-                        },
-                    )
-                }
-                GapPolicy::Includes => {
-                    let outer_span = top_max - bottom_min;
-                    let combined_height = top.half_size.y * 2.0 + bottom.half_size.y * 2.0;
-                    if outer_span <= 0.0 || combined_height <= 0.0 {
-                        return None;
-                    }
-
-                    let scale = combined_height / outer_span;
-                    if scale <= f32::EPSILON {
-                        return None;
-                    }
-
-                    let target_top_height = (top.half_size.y * 2.0) / scale;
-                    let target_bottom_height = (bottom.half_size.y * 2.0) / scale;
-                    (
-                        BattleBoxBoundary {
-                            half_size: Vec2::new(top.half_size.x, target_top_height * 0.5),
-                            center: Vec2::new(top.center.x, top_max - target_top_height * 0.5),
-                        },
-                        BattleBoxBoundary {
-                            half_size: Vec2::new(bottom.half_size.x, target_bottom_height * 0.5),
-                            center: Vec2::new(
-                                bottom.center.x,
-                                bottom_min + target_bottom_height * 0.5,
-                            ),
-                        },
-                    )
-                }
-            };
-
-            let merged_boundary = merge_boundaries(&target_top, &target_bottom);
-            Some(MergeAnimationPlan {
-                axis,
-                ordered_indices,
-                target_boundary_a: target_top,
-                target_boundary_b: target_bottom,
-                merged_boundary,
-            })
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn approx_eq(a: f32, b: f32) {
-        assert!(
-            (a - b).abs() < 0.001,
-            "expected {a} ~= {b}, diff={}",
-            (a - b).abs()
-        );
-    }
-
-    fn boundary(width: f32, height: f32) -> BattleBoxBoundary {
-        BattleBoxBoundary {
-            center: Vec2::ZERO,
-            half_size: Vec2::new(width / 2.0, height / 2.0),
-        }
-    }
-
-    #[test]
-    fn horizontal_split_positive_position_makes_top_smaller() {
-        let original = boundary(300.0, 200.0);
-        let (top, bottom) = split_rect_box(
-            &original,
-            &SplitAxis::Horizontal,
-            20.0,
-            0.0,
-            GapPolicy::Expands,
-        );
-
-        approx_eq(top.half_size.y * 2.0, 80.0);
-        approx_eq(bottom.half_size.y * 2.0, 120.0);
-        assert!(top.center.y > bottom.center.y);
-    }
-
-    #[test]
-    fn horizontal_includes_preserves_outer_edges_and_gap() {
-        let original = boundary(300.0, 200.0);
-        let gap = 20.0;
-        let (top, bottom) = split_rect_box(
-            &original,
-            &SplitAxis::Horizontal,
-            20.0,
-            gap,
-            GapPolicy::Includes,
-        );
-
-        approx_eq(top.center.y + top.half_size.y, 100.0);
-        approx_eq(bottom.center.y - bottom.half_size.y, -100.0);
-        approx_eq(
-            (top.center.y - top.half_size.y) - (bottom.center.y + bottom.half_size.y),
-            gap,
-        );
-    }
-
-    #[test]
-    fn vertical_includes_preserves_outer_edges_and_gap() {
-        let original = boundary(300.0, 200.0);
-        let gap = 20.0;
-        let (left, right) = split_rect_box(
-            &original,
-            &SplitAxis::Vertical,
-            20.0,
-            gap,
-            GapPolicy::Includes,
-        );
-
-        approx_eq(left.center.x - left.half_size.x, -150.0);
-        approx_eq(right.center.x + right.half_size.x, 150.0);
-        approx_eq(
-            (right.center.x - right.half_size.x) - (left.center.x + left.half_size.x),
-            gap,
-        );
-    }
-
-    #[test]
-    fn structured_vertical_split_gap_matches_requested_visible_gap() {
-        let original = boundary(300.0, 200.0);
-        let style = BattleBoxVisualStyle {
-            border_width: 5.0,
-            fill_shader: None,
-            structure_file: Some("shared/view_structures/view_box.sdf.ron".to_string()),
-            fill_color: Color::BLACK,
-        };
-        let requested_visible_gap = 20.0;
-        let (left, right) = split_rect_box(
-            &original,
-            &SplitAxis::Vertical,
-            0.0,
-            style.boundary_gap_for_visible_gap(requested_visible_gap),
-            GapPolicy::Expands,
-        );
-
-        let visible_gap = (right.center.x - right.half_size.x - style.border_width)
-            - (left.center.x + left.half_size.x + style.border_width);
-        approx_eq(visible_gap, requested_visible_gap);
-    }
-
-    #[test]
-    fn structured_zero_visible_gap_keeps_outer_edges_touching() {
-        let original = boundary(300.0, 200.0);
-        let style = BattleBoxVisualStyle {
-            border_width: 5.0,
-            fill_shader: None,
-            structure_file: Some("shared/view_structures/view_box.sdf.ron".to_string()),
-            fill_color: Color::BLACK,
-        };
-        let (top, bottom) = split_rect_box(
-            &original,
-            &SplitAxis::Horizontal,
-            0.0,
-            style.boundary_gap_for_visible_gap(0.0),
-            GapPolicy::Includes,
-        );
-
-        let visible_gap = (top.center.y - top.half_size.y - style.border_width)
-            - (bottom.center.y + bottom.half_size.y + style.border_width);
-        approx_eq(visible_gap, 0.0);
-    }
-
-    #[test]
-    fn vertical_expands_merge_recovers_original_boundary() {
-        let original = boundary(300.0, 200.0);
-        let (left, right) = split_rect_box(
-            &original,
-            &SplitAxis::Vertical,
-            20.0,
-            30.0,
-            GapPolicy::Expands,
-        );
-        let expected = split_rect_box(
-            &original,
-            &SplitAxis::Vertical,
-            20.0,
-            0.0,
-            GapPolicy::Expands,
-        );
-
-        let plan =
-            plan_merge_animation(&[left.clone(), right.clone()], GapPolicy::Expands).unwrap();
-
-        assert_eq!(plan.axis, SplitAxis::Vertical);
-        approx_eq(plan.merged_boundary.center.x, original.center.x);
-        approx_eq(plan.merged_boundary.center.y, original.center.y);
-        approx_eq(plan.merged_boundary.half_size.x, original.half_size.x);
-        approx_eq(plan.merged_boundary.half_size.y, original.half_size.y);
-        approx_eq(plan.target_boundary_a.center.x, expected.0.center.x);
-        approx_eq(plan.target_boundary_b.center.x, expected.1.center.x);
-        approx_eq(plan.target_boundary_a.half_size.x, expected.0.half_size.x);
-        approx_eq(plan.target_boundary_b.half_size.x, expected.1.half_size.x);
-    }
-
-    #[test]
-    fn horizontal_includes_merge_recovers_original_boundary() {
-        let original = boundary(300.0, 200.0);
-        let (top, bottom) = split_rect_box(
-            &original,
-            &SplitAxis::Horizontal,
-            20.0,
-            20.0,
-            GapPolicy::Includes,
-        );
-        let expected = split_rect_box(
-            &original,
-            &SplitAxis::Horizontal,
-            20.0,
-            0.0,
-            GapPolicy::Includes,
-        );
-
-        let plan =
-            plan_merge_animation(&[top.clone(), bottom.clone()], GapPolicy::Includes).unwrap();
-
-        assert_eq!(plan.axis, SplitAxis::Horizontal);
-        approx_eq(plan.merged_boundary.center.x, original.center.x);
-        approx_eq(plan.merged_boundary.center.y, original.center.y);
-        approx_eq(plan.merged_boundary.half_size.x, original.half_size.x);
-        approx_eq(plan.merged_boundary.half_size.y, original.half_size.y);
-        approx_eq(plan.target_boundary_a.center.y, expected.0.center.y);
-        approx_eq(plan.target_boundary_b.center.y, expected.1.center.y);
-        approx_eq(plan.target_boundary_a.half_size.y, expected.0.half_size.y);
-        approx_eq(plan.target_boundary_b.half_size.y, expected.1.half_size.y);
-    }
-
-    #[test]
-    fn player_rebind_prefers_box_that_currently_contains_player() {
-        let original = boundary(300.0, 200.0);
-        let (left, right) = split_rect_box(
-            &original,
-            &SplitAxis::Vertical,
-            100.0,
-            0.0,
-            GapPolicy::Expands,
-        );
-        let player_pos = Vec2::new(70.0, 0.0);
-        let collider = PhysicsCollider::Circle { radius: 8.0 };
-
-        assert_eq!(
-            select_box_id_for_player(player_pos, &collider, &left, &right, "left", "right"),
-            "left"
-        );
-    }
-
-    #[test]
-    fn dynamic_box_choice_switches_when_current_binding_no_longer_contains_player() {
-        let original = boundary(300.0, 200.0);
-        let (left, right) = split_rect_box(
-            &original,
-            &SplitAxis::Vertical,
-            100.0,
-            0.0,
-            GapPolicy::Expands,
-        );
-        let player_pos = Vec2::new(70.0, 0.0);
-        let collider = PhysicsCollider::Circle { radius: 8.0 };
-        let candidates = vec![
-            ("left".to_string(), left.clone()),
-            ("right".to_string(), right.clone()),
-        ];
-
-        assert_eq!(
-            choose_box_index_for_player(Some("right"), player_pos, &collider, &candidates),
-            Some(0)
-        );
-    }
-}
-
-fn signed_distance_to_box_with_collider(
-    boundary: &BattleBoxBoundary,
-    player_pos: Vec2,
-    collider: &PhysicsCollider,
-) -> f32 {
-    let collider_half_size = match collider {
-        PhysicsCollider::Circle { radius } => Vec2::splat(*radius),
-        PhysicsCollider::Box { half_size } => *half_size,
-    };
-    let effective_half_size = (boundary.half_size - collider_half_size).max(Vec2::ZERO);
-    BattleBoxBoundary {
-        half_size: effective_half_size,
-        center: boundary.center,
-    }
-    .sdf_distance(player_pos)
-}
-
-/// Determine which of two boxes a player should be rebound to.
-fn select_box_id_for_player(
-    player_pos: Vec2,
-    collider: &PhysicsCollider,
-    box_a: &BattleBoxBoundary,
-    box_b: &BattleBoxBoundary,
-    id_a: &str,
-    id_b: &str,
-) -> String {
-    let dist_a = signed_distance_to_box_with_collider(box_a, player_pos, collider);
-    let dist_b = signed_distance_to_box_with_collider(box_b, player_pos, collider);
-
-    if dist_a <= 0.0 && dist_b > 0.0 {
-        id_a.to_string()
-    } else if dist_b <= 0.0 && dist_a > 0.0 {
-        id_b.to_string()
-    } else if dist_a <= dist_b {
-        id_a.to_string()
-    } else {
-        id_b.to_string()
-    }
-}
-
-fn choose_box_index_for_player(
-    current_bound: Option<&str>,
-    player_pos: Vec2,
-    collider: &PhysicsCollider,
-    candidates: &[(String, BattleBoxBoundary)],
-) -> Option<usize> {
-    if candidates.is_empty() {
-        return None;
-    }
-
-    if let Some(current_id) = current_bound {
-        if let Some((index, _)) = candidates.iter().enumerate().find(|(_, (id, boundary))| {
-            id == current_id
-                && signed_distance_to_box_with_collider(boundary, player_pos, collider) <= 0.0
-        }) {
-            return Some(index);
-        }
-    }
-
-    if let Some((index, _)) = candidates.iter().enumerate().find(|(_, (_, boundary))| {
-        signed_distance_to_box_with_collider(boundary, player_pos, collider) <= 0.0
-    }) {
-        return Some(index);
-    }
-
-    candidates
-        .iter()
-        .enumerate()
-        .min_by(|(_, (_, a)), (_, (_, b))| {
-            signed_distance_to_box_with_collider(a, player_pos, collider).total_cmp(
-                &signed_distance_to_box_with_collider(b, player_pos, collider),
-            )
-        })
-        .map(|(index, _)| index)
-}
-
-// ─── Helper: resolve boundary from box entity ───────────────────────
-
-/// Resolve the `BattleBoxBoundary` from a battle box entity.
-/// Returns `None` if the box is inactive or collision is disabled.
-fn resolve_boundary(
-    transform: &GlobalTransform,
-    ui_box: Option<&ViewBox>,
-    am_bounds: Option<&AlightMotionBattleBoxBounds>,
-    state: &BattleBoxState,
-) -> Option<BattleBoxBoundary> {
-    if !state.active || !state.collision_enabled {
-        return None;
-    }
-    if let Some(vb) = ui_box {
-        Some(BattleBoxBoundary::from_ui_box(
-            vb.width(),
-            vb.height(),
-            transform.translation().truncate(),
-        ))
-    } else if let Some(am) = am_bounds {
-        let center = transform.translation().truncate() + am.center_offset;
-        Some(BattleBoxBoundary::from_ui_box(am.width, am.height, center))
-    } else {
-        None
-    }
-}
-
-fn resolve_visual_style(
-    ui_box: Option<&ViewBox>,
-    style: Option<&BattleBoxVisualStyle>,
-) -> BattleBoxVisualStyle {
-    style
-        .cloned()
-        .or_else(|| ui_box.map(BattleBoxVisualStyle::from_view_box))
-        .unwrap_or_default()
-}
-
-fn collect_battle_box_candidates(
-    target_id: &str,
-    ui_boxes: &UiBattleBoxReadQuery,
-    am_boxes: &AmBattleBoxReadQuery,
-) -> Vec<BattleBoxCandidate> {
-    let mut candidates = Vec::new();
-
-    for (entity, tf, vb, box_id, state, style) in ui_boxes.iter() {
-        if box_id.0 != target_id {
-            continue;
-        }
-
-        candidates.push(BattleBoxCandidate {
-            entity,
-            id: box_id.0.clone(),
-            kind: BattleBoxSourceKind::Ui,
-            active: state.active,
-            collision_enabled: state.collision_enabled,
-            boundary: resolve_boundary(tf, Some(vb), None, state),
-            visual_style: resolve_visual_style(Some(vb), style),
-        });
-    }
-
-    for (entity, tf, am, box_id, state, style) in am_boxes.iter() {
-        if box_id.0 != target_id {
-            continue;
-        }
-
-        candidates.push(BattleBoxCandidate {
-            entity,
-            id: box_id.0.clone(),
-            kind: BattleBoxSourceKind::Am,
-            active: state.active,
-            collision_enabled: state.collision_enabled,
-            boundary: resolve_boundary(tf, None, Some(am), state),
-            visual_style: resolve_visual_style(None, style),
-        });
-    }
-
-    candidates
-}
-
-fn describe_battle_box_candidates(candidates: &[BattleBoxCandidate]) -> String {
-    candidates
-        .iter()
-        .map(BattleBoxCandidate::summary)
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-fn resolve_live_battle_box(
-    op_name: &str,
-    target_id: &str,
-    ui_boxes: &UiBattleBoxReadQuery,
-    am_boxes: &AmBattleBoxReadQuery,
-) -> Option<BattleBoxCandidate> {
-    let candidates = collect_battle_box_candidates(target_id, ui_boxes, am_boxes);
-    if candidates.is_empty() {
-        warn!("{op_name}: source box '{target_id}' not found");
-        return None;
-    }
-
-    let live = candidates
-        .iter()
-        .filter(|candidate| candidate.is_live())
-        .cloned()
-        .collect::<Vec<_>>();
-
-    match live.len() {
-        1 => live.into_iter().next(),
-        0 => {
-            warn!(
-                "{op_name}: source box '{target_id}' exists but has no usable live match: {}",
-                describe_battle_box_candidates(&candidates)
-            );
-            None
-        }
-        count => {
-            warn!(
-                "{op_name}: source box '{target_id}' is ambiguous; found {count} live matches: {}",
-                describe_battle_box_candidates(&live)
-            );
-            None
-        }
-    }
-}
-
-fn hide_entity_and_descendants(
-    commands: &mut Commands,
-    entity: Entity,
-    children_query: &Query<&Children>,
-) {
-    commands.entity(entity).insert(Visibility::Hidden);
-
-    let Ok(children) = children_query.get(entity) else {
-        return;
-    };
-
-    for child in children.iter() {
-        hide_entity_and_descendants(commands, child, children_query);
-    }
-}
-
-fn retire_battle_box(commands: &mut Commands, entity: Entity, children_query: &Query<&Children>) {
-    hide_entity_and_descendants(commands, entity, children_query);
-    commands
-        .entity(entity)
-        .remove::<(BattleBox, BattleBoxId, BattleBoxState, BattleBoxVisualStyle)>();
-}
-
-fn retire_existing_battle_boxes_with_id(
-    commands: &mut Commands,
-    op_name: &str,
-    target_id: &str,
-    keep_entities: &[Entity],
-    children_query: &Query<&Children>,
-    ui_boxes: &UiBattleBoxReadQuery,
-    am_boxes: &AmBattleBoxReadQuery,
-) {
-    for candidate in collect_battle_box_candidates(target_id, ui_boxes, am_boxes) {
-        if keep_entities.contains(&candidate.entity) {
-            continue;
-        }
-
-        warn!(
-            "{op_name}: retiring pre-existing box for result id '{target_id}': {}",
-            candidate.summary()
-        );
-        retire_battle_box(commands, candidate.entity, children_query);
-    }
 }
 
 // ─── Systems ────────────────────────────────────────────────────────
@@ -1430,23 +700,19 @@ fn handle_merge_battle_boxes_system(
                     ev.source_boxes.0, ev.source_boxes.1
                 );
                 let merged = merge_boundaries(&boundaries[0], &boundaries[1]);
-                for entity in &source_entities {
-                    retire_battle_box(&mut commands, *entity, &children_query);
-                }
-                spawn_standalone_box(
+                finalize_merged_battle_box(
                     &mut commands,
+                    &source_entities,
+                    &children_query,
                     &mut meshes,
                     &mut sdf_materials,
                     &mut color_materials,
+                    &mut player_query,
+                    &ev.source_boxes,
                     &ev.result_box,
                     &merged,
                     &style,
                 );
-                for mut bound in player_query.iter_mut() {
-                    if bound.0 == ev.source_boxes.0 || bound.0 == ev.source_boxes.1 {
-                        bound.0 = ev.result_box.clone();
-                    }
-                }
                 continue;
             };
 
@@ -1479,23 +745,19 @@ fn handle_merge_battle_boxes_system(
                 .map(|plan| plan.merged_boundary.clone())
                 .unwrap_or_else(|| merge_boundaries(&boundaries[0], &boundaries[1]));
 
-            for entity in &source_entities {
-                retire_battle_box(&mut commands, *entity, &children_query);
-            }
-            spawn_standalone_box(
+            finalize_merged_battle_box(
                 &mut commands,
+                &source_entities,
+                &children_query,
                 &mut meshes,
                 &mut sdf_materials,
                 &mut color_materials,
+                &mut player_query,
+                &ev.source_boxes,
                 &ev.result_box,
                 &merged,
                 &style,
             );
-            for mut bound in player_query.iter_mut() {
-                if bound.0 == ev.source_boxes.0 || bound.0 == ev.source_boxes.1 {
-                    bound.0 = ev.result_box.clone();
-                }
-            }
         }
 
         info!(
@@ -1511,342 +773,5 @@ fn handle_merge_battle_boxes_system(
             ev.duration,
             ev.easing
         );
-    }
-}
-
-// ─── Standalone Box Helpers ─────────────────────────────────────────
-
-/// Spawn a standalone battle box entity with its own SDF visual.
-fn spawn_standalone_box(
-    commands: &mut Commands,
-    meshes: &mut ResMut<Assets<Mesh>>,
-    sdf_materials: &mut ResMut<Assets<SdfMaterial>>,
-    color_materials: &mut ResMut<Assets<ColorMaterial>>,
-    id: &str,
-    boundary: &BattleBoxBoundary,
-    visual_style: &BattleBoxVisualStyle,
-) {
-    let view_box = visual_style.to_view_box(boundary.half_size.x * 2.0, boundary.half_size.y * 2.0);
-    let entity = commands
-        .spawn((
-            BattleBox,
-            BattleBoxId(id.to_string()),
-            BattleBoxState::default(),
-            visual_style.clone(),
-            AlightMotionBattleBoxBounds {
-                width: boundary.half_size.x * 2.0,
-                height: boundary.half_size.y * 2.0,
-                center_offset: Vec2::ZERO,
-            },
-            Transform::from_translation(boundary.center.extend(0.0)),
-            GlobalTransform::default(),
-            Visibility::default(),
-            InheritedVisibility::default(),
-            ViewVisibility::default(),
-            Name::new(format!("BattleBox:{id}")),
-        ))
-        .id();
-
-    spawn_view_box_sdf_children(
-        commands,
-        entity,
-        &view_box,
-        meshes,
-        sdf_materials,
-        color_materials,
-    );
-}
-
-/// Merge two boundaries into one AABB that encloses both.
-fn merge_boundaries(a: &BattleBoxBoundary, b: &BattleBoxBoundary) -> BattleBoxBoundary {
-    let min = Vec2::new(
-        (a.center.x - a.half_size.x).min(b.center.x - b.half_size.x),
-        (a.center.y - a.half_size.y).min(b.center.y - b.half_size.y),
-    );
-    let max = Vec2::new(
-        (a.center.x + a.half_size.x).max(b.center.x + b.half_size.x),
-        (a.center.y + a.half_size.y).max(b.center.y + b.half_size.y),
-    );
-    let center = (min + max) / 2.0;
-    let half_size = (max - min) / 2.0;
-    BattleBoxBoundary { half_size, center }
-}
-
-/// Spawn a standalone battle box entity without SDF children.
-/// Used for animation where visual is added separately.
-fn spawn_standalone_box_entity(
-    commands: &mut Commands,
-    id: &str,
-    boundary: &BattleBoxBoundary,
-    visual_style: &BattleBoxVisualStyle,
-) -> Entity {
-    commands
-        .spawn((
-            BattleBox,
-            BattleBoxId(id.to_string()),
-            BattleBoxState::default(),
-            visual_style.clone(),
-            AlightMotionBattleBoxBounds {
-                width: boundary.half_size.x * 2.0,
-                height: boundary.half_size.y * 2.0,
-                center_offset: Vec2::ZERO,
-            },
-            Transform::from_translation(boundary.center.extend(0.0)),
-            GlobalTransform::default(),
-            Visibility::default(),
-            InheritedVisibility::default(),
-            ViewVisibility::default(),
-            Name::new(format!("BattleBox:{id}")),
-        ))
-        .id()
-}
-
-// ─── Animation Systems ────────────────────────────────────────────────
-
-/// System to animate battle box split animations.
-/// 战斗框分裂动画系统。
-fn animate_battle_box_split_system(
-    mut commands: Commands,
-    time: Res<Time>,
-    mut animations: Query<(Entity, &mut BattleBoxSplitAnimation)>,
-    mut box_query: Query<(&mut Transform, &mut AlightMotionBattleBoxBounds)>,
-    child_query: Query<&Children>,
-    mut shape_query: Query<(
-        &mut crate::core::view::sdf_shape::ViewSdfShape,
-        &MeshMaterial2d<SdfMaterial>,
-        &mut Mesh2d,
-    )>,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut sdf_materials: ResMut<Assets<SdfMaterial>>,
-) {
-    for (anim_entity, mut anim) in animations.iter_mut() {
-        anim.progress += time.delta_secs() / anim.duration;
-        let t = anim.progress.min(1.0);
-
-        // Easing: smooth step (ease in-out)
-        let eased = anim.easing.sample(t);
-        let current_visible_gap = anim.target_visible_gap * eased;
-        let current_gap = anim
-            .visual_style
-            .boundary_gap_for_visible_gap(current_visible_gap);
-        let (current_a, current_b) = split_rect_box(
-            &anim.original_boundary,
-            &anim.split_axis,
-            anim.split_position,
-            current_gap,
-            anim.gap_policy,
-        );
-
-        if anim.progress >= 1.0 {
-            apply_boundary_to_box(
-                anim.box_entity_a,
-                &current_a,
-                &anim.visual_style,
-                &mut box_query,
-                &child_query,
-                &mut shape_query,
-                &mut meshes,
-                &mut sdf_materials,
-            );
-            apply_boundary_to_box(
-                anim.box_entity_b,
-                &current_b,
-                &anim.visual_style,
-                &mut box_query,
-                &child_query,
-                &mut shape_query,
-                &mut meshes,
-                &mut sdf_materials,
-            );
-            commands.entity(anim_entity).despawn();
-            continue;
-        }
-
-        apply_boundary_to_box(
-            anim.box_entity_a,
-            &current_a,
-            &anim.visual_style,
-            &mut box_query,
-            &child_query,
-            &mut shape_query,
-            &mut meshes,
-            &mut sdf_materials,
-        );
-        apply_boundary_to_box(
-            anim.box_entity_b,
-            &current_b,
-            &anim.visual_style,
-            &mut box_query,
-            &child_query,
-            &mut shape_query,
-            &mut meshes,
-            &mut sdf_materials,
-        );
-    }
-}
-
-/// System to animate battle box merge animations.
-/// 战斗框合并动画系统。
-fn animate_battle_box_merge_system(
-    mut commands: Commands,
-    time: Res<Time>,
-    mut animations: Query<(Entity, &mut BattleBoxMergeAnimation)>,
-    mut player_query: Query<&mut BoundToBattleBox, With<BehaviorParams>>,
-    mut box_query: Query<(&mut Transform, &mut AlightMotionBattleBoxBounds)>,
-    child_query: Query<&Children>,
-    mut shape_query: Query<(
-        &mut crate::core::view::sdf_shape::ViewSdfShape,
-        &MeshMaterial2d<SdfMaterial>,
-        &mut Mesh2d,
-    )>,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut sdf_materials: ResMut<Assets<SdfMaterial>>,
-    mut color_materials: ResMut<Assets<ColorMaterial>>,
-) {
-    for (anim_entity, mut anim) in animations.iter_mut() {
-        anim.progress += time.delta_secs() / anim.duration;
-        let t = anim.progress.min(1.0);
-        let eased = anim.easing.sample(t);
-
-        let current_a = lerp_boundary(&anim.start_boundary_a, &anim.target_boundary_a, eased);
-        let current_b = lerp_boundary(&anim.start_boundary_b, &anim.target_boundary_b, eased);
-
-        apply_boundary_to_box(
-            anim.box_entity_a,
-            &current_a,
-            &anim.visual_style,
-            &mut box_query,
-            &child_query,
-            &mut shape_query,
-            &mut meshes,
-            &mut sdf_materials,
-        );
-        apply_boundary_to_box(
-            anim.box_entity_b,
-            &current_b,
-            &anim.visual_style,
-            &mut box_query,
-            &child_query,
-            &mut shape_query,
-            &mut meshes,
-            &mut sdf_materials,
-        );
-
-        if anim.progress < 1.0 {
-            continue;
-        }
-
-        retire_battle_box(&mut commands, anim.box_entity_a, &child_query);
-        retire_battle_box(&mut commands, anim.box_entity_b, &child_query);
-
-        spawn_standalone_box(
-            &mut commands,
-            &mut meshes,
-            &mut sdf_materials,
-            &mut color_materials,
-            &anim.result_box,
-            &anim.merged_boundary,
-            &anim.visual_style,
-        );
-
-        for mut bound in player_query.iter_mut() {
-            if bound.0 == anim.source_boxes.0 || bound.0 == anim.source_boxes.1 {
-                bound.0 = anim.result_box.clone();
-            }
-        }
-
-        commands.entity(anim_entity).despawn();
-    }
-}
-
-fn apply_boundary_to_box(
-    box_entity: Entity,
-    boundary: &BattleBoxBoundary,
-    visual_style: &BattleBoxVisualStyle,
-    box_query: &mut Query<(&mut Transform, &mut AlightMotionBattleBoxBounds)>,
-    child_query: &Query<&Children>,
-    shape_query: &mut Query<(
-        &mut crate::core::view::sdf_shape::ViewSdfShape,
-        &MeshMaterial2d<SdfMaterial>,
-        &mut Mesh2d,
-    )>,
-    meshes: &mut ResMut<Assets<Mesh>>,
-    sdf_materials: &mut ResMut<Assets<SdfMaterial>>,
-) {
-    if let Ok((mut transform, mut bounds)) = box_query.get_mut(box_entity) {
-        transform.translation = boundary.center.extend(0.0);
-        bounds.width = boundary.half_size.x * 2.0;
-        bounds.height = boundary.half_size.y * 2.0;
-        update_sdf_visual(
-            &box_entity,
-            boundary,
-            visual_style,
-            child_query,
-            shape_query,
-            meshes,
-            sdf_materials,
-        );
-    }
-}
-
-/// Update SDF visual for a battle box entity during animation.
-fn update_sdf_visual(
-    box_entity: &Entity,
-    boundary: &BattleBoxBoundary,
-    visual_style: &BattleBoxVisualStyle,
-    child_query: &Query<&Children>,
-    shape_query: &mut Query<(
-        &mut crate::core::view::sdf_shape::ViewSdfShape,
-        &MeshMaterial2d<SdfMaterial>,
-        &mut Mesh2d,
-    )>,
-    meshes: &mut ResMut<Assets<Mesh>>,
-    sdf_materials: &mut ResMut<Assets<SdfMaterial>>,
-) {
-    let Ok(children) = child_query.get(*box_entity) else {
-        return;
-    };
-
-    let mut queue: VecDeque<Entity> = VecDeque::from(children.to_vec());
-    let mut sdf_entities = Vec::new();
-    let expected_shapes = if visual_style.structure_file.is_some() {
-        2
-    } else {
-        1
-    };
-
-    while let Some(entity) = queue.pop_front() {
-        if shape_query.get(entity).is_ok() {
-            sdf_entities.push(entity);
-            if sdf_entities.len() >= expected_shapes {
-                break;
-            }
-        }
-        if let Ok(grandchildren) = child_query.get(entity) {
-            queue.extend(grandchildren.to_vec());
-        }
-    }
-
-    let box_half_width = boundary.half_size.x;
-    let box_half_height = boundary.half_size.y;
-
-    for (index, entity) in sdf_entities.into_iter().enumerate() {
-        let (half_width, half_height) = if expected_shapes == 1 || index > 0 {
-            (box_half_width, box_half_height)
-        } else {
-            (
-                box_half_width + visual_style.border_width,
-                box_half_height + visual_style.border_width,
-            )
-        };
-
-        if let Ok((mut shape, material_handle, mut mesh_handle)) = shape_query.get_mut(entity) {
-            shape.half_width = half_width;
-            shape.half_height = half_height;
-            if let Some(material) = sdf_materials.get_mut(&material_handle.0) {
-                *material = shape.to_material();
-            }
-            mesh_handle.0 = meshes.add(shape.create_mesh());
-        }
     }
 }
